Extracting a RankedInfo component
